DockPipe and CablePipe cleanup and Fixes + Rename
