+++ Content.Server/NodeContainer/Nodes/PipeNode.cs
@@ -163,6 +163,11 @@ namespace Content.Server.NodeContainer.Nodes
             MapGridComponent? grid,
             IEntityManager entMan)
         {
+            // Starlight Start: Moved to before if alwaysReachable
+            if (!xform.Anchored || grid == null)
+                yield break;
+            // Starlight End: Moved to before if alwaysReachable
+
             if (_alwaysReachable != null)
             {
                 var remQ = new RemQueue<PipeNode>();
@@ -181,8 +186,10 @@ namespace Content.Server.NodeContainer.Nodes
                 }
             }
 
-            if (!xform.Anchored || grid == null)
-                yield break;
+            // Starlight edit Start: Moved to before if alwaysReachable
+            // if (!xform.Anchored || grid == null)
+            //     yield break;
+            // Starlight edit End: Moved to before if alwaysReachable
 
             var pos = grid.TileIndicesFor(xform.Coordinates);
